Finished up raytracer.

+++ RayTracer/Source/Framework/HittableList.cpp
@@ -6,7 +6,7 @@ bool HittableList::Hit(const Ray& r, float tMin, float tMax, HitRecord& rec) con
 	double closestSoFar = tMax;
 	for (int i = 0; i < listSize; i++)
 	{
-		if (list[i]->Hit(r, tMin, closestSoFar, tempRec)) {
+		if (list[i]->Hit(r, tMin, static_cast<float>(closestSoFar), tempRec)) {
 			hitAnything = true;
 			closestSoFar = tempRec.t;
 			rec = tempRec;
